Método getById para Acudiente

+++ src/main/java/com/app/akdemy/repository/AcudienteRepository.java
@@ -4,6 +4,6 @@ import com.app.akdemy.entity.Acudiente;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AcudienteRepository extends CrudRepository<Acudiente, Integer>{
+public interface AcudienteRepository extends CrudRepository<Acudiente, Long>{
     
 }
